requirements.txt
fastapi
uvicorn
google-genai
ddgs
python-dotenv
slowapi
httpx
pydantic
python-multipart
markdown
